requirements.txt
streamlit>=1.31.0
chromadb>=0.4.22
google-generativeai>=0.3.2
python-dotenv>=1.0.0
httpx>=0.26.0
aiohttp>=3.9.0
python-magic>=0.4.27
pandas>=2.2.0
numpy>=1.26.0
langchain>=0.1.0
streamlit-chat>=0.1.1
langchainhub>=0.1.14
google-cloud-aiplatform>=1.42.0
email-validator>=2.1.0
pydantic>=2.6.0
aiofiles>=23.2.1
async-timeout>=4.0.3
mimetypes-magic>=0.1.0